hide test:e2e for contact-us

+++ client/tests/e2e/specs/candidatContactUs.js
@@ -1,5 +1,5 @@
 describe('Contact Us', () => {
-  if (Cypress.env('VUE_APP_CLIENT_BUILD_INFO') !== 'COVID') {
+  if (Cypress.env('VUE_APP_CLIENT_BUILD_INFO') !== 'COVID' && Cypress.env('VUE_APP_CLIENT_BUILD_WITH_CONTACT_US') === true) {
     before(() => {
       cy.deleteAllMails()
       cy.candidatConnection(Cypress.env('emailCandidatContactUs'))
